cortex_on/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.7.0
asyncio-atexit==1.0.1
attrs==24.3.0
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cobble==0.1.4
colorama==0.4.6
cryptography==44.0.0
Deprecated==1.2.15
distro==1.9.0
docker==7.1.0
eval_type_backport==0.2.2
executing==2.1.0
fastapi==0.115.6
frozenlist==1.5.0
google-auth==2.37.0
googleapis-common-protos==1.66.0
griffe==1.5.4
groq==0.15.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
jiter==0.8.2
jsonpath-python==1.0.6
logfire==2.11.0
logfire-api==2.11.0
lxml==5.3.0
mammoth==1.8.0
markdown-it-py==3.0.0
markdownify==0.14.1
mdurl==0.1.2
mistralai==1.2.5
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.2.1
openai==1.74.0
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
packaging==24.2
pandas==2.2.3
pathvalidate==3.2.1
pdfminer.six==20240706
pillow==11.0.0
propcache==0.2.1
protobuf==5.29.2
puremagic==1.28
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.4
pydantic-ai==0.1.2
pydantic-ai-slim==0.1.2
pydantic_core==2.27.2
pydantic-ai==0.1.2
pydantic-ai-slim==0.1.2
mcp==1.6.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-pptx==1.0.2
pytz==2024.2
requests==2.32.3
rich==13.9.4
rsa==4.9
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.3
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
websockets==14.1
wrapt==1.17.0
XlsxWriter==3.2.0
yarl==1.18.3
zipp==3.21.0
fast-graphrag==0.0.4
llama_parse==0.5.19
mcp-server-time